rename package `awsecret` to `aws`
This commit renames the `awsecret` package
to make the API more idiomatic.

Before, `awsecret` should only contain types
and functions related to the AWS SecretsManager.
Therefore, it contained the generic `KeyStore` type.

Now, the package `aws` groups all AWS services that
can be used as key store. So, there is a `aws.SecretsManager`
type that implements a key store using the AWS SecretsManager
as storage.
